refactor: global improvements
