+++ Cargo.toml
@@ -24,14 +24,16 @@ edition = "2021"
 [workspace.dependencies]
 # Core IC dependencies
 candid = "0.10"
+ic-agent = "0.39"
 ic-cdk = "0.16"
-ic-types = "0.7"
 ic-canister-log = "0.2"
 ic-canisters-http-types = { git = "https://github.com/dfinity/ic", package = "ic-canisters-http-types" }
 ic-crypto-ed25519 = { git = "https://github.com/dfinity/ic", package = "ic-crypto-ed25519" }
 ic-management-canister-types = { git = "https://github.com/dfinity/ic", package = "ic-management-canister-types" }
 ic-metrics-encoder = "1.1"
 ic-stable-structures = "0.6.5"
+ic-types = "0.7"
+ic-utils = "0.39"
 
 # Project-specific dependencies
 ic-solana = { path = "src/ic-solana" }
