+++ src/ic-solana-provider/Cargo.toml
@@ -36,6 +36,6 @@ solana-client = "2.0"
 ctor = "0.2.8"
 lazy_static = "1.5"
 tokio = { version = "1.0", features = ["rt-multi-thread", "macros"] }
-ic-agent = "0.38.1"
-ic-utils = "0.38.1"
+ic-agent = { workspace = true, features = ["ring"] }
+ic-utils = { workspace = true }
 ring = "0.17.8"
