+++ src/ic-solana-wallet/Cargo.toml
@@ -25,3 +25,19 @@ serde_json = { workspace = true }
 serde_bytes = { workspace = true }
 url = "2.5.0"
 hex = "0.4.3"
+bs58 = "0.5.1"
+
+[dev-dependencies]
+anyhow = "1.0.40"
+bincode = "1.3.3"
+flate2 = "1.0.24"
+pocket-ic = { workspace = true }
+rand = "0.8.5"
+solana-sdk = "2.0"
+solana-client = "2.0"
+ctor = "0.2.8"
+lazy_static = "1.5"
+tokio = { version = "1.0", features = ["rt-multi-thread", "macros"] }
+ic-agent = { workspace = true, features = ["ring"] }
+ic-utils = { workspace = true }
+ring = "0.17.8"
