+++ src/ic-solana-wallet/src/state.rs
@@ -5,7 +5,7 @@ use {
 };
 
 thread_local! {
-    pub static STATE: RefCell<Option<State>> = RefCell::new(None);
+    pub static STATE: RefCell<Option<State>> = const { RefCell::new(None) };
 }
 
 /// Solana RPC canister initialization data.
