--- src/ic-solana-wallet/src/types.rs
@@ -1,4 +1,0 @@
-use {
-    candid::CandidType,
-    serde::{Deserialize, Serialize},
-};
