+++ src/ic-solana/src/lib.rs
@@ -2,7 +2,6 @@ pub mod constants;
 pub mod request;
 pub mod response;
 pub mod rpc_client;
-pub mod rpc_result;
 pub mod types;
 pub mod utils;
